02 - Simple RAG/pyproject.toml
[project]
name = "session02"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.36.0",
    "gradio>=5.45.0",
    "openai>=1.107.1",
    "pinecone>=7.3.0",
    "python-dotenv>=1.1.1",
]
